% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{any_satisfied}
\alias{any_satisfied}
\title{Checks if any one of a set of assertions is satisfied}
\usage{
any_satisfied(...)
}
\arguments{
\item{...}{the various assertions that can be satisfied}
}
\description{
Allows to check for multiple cases, and pass the test if any of the cases
is satisfied.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(
  qscheck::any_satisfied(
   qscheck::is_positive_integer_value(value),
   qscheck::is_list(value)
  )
)
# For check
if (
  qscheck::any_satisfied(
   qscheck::is_positive_integer_value(value),
   qscheck::is_list(value)
  )
) {}
}

}
\concept{combined}
