% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Shorter version of assertthat::assert_that}
\usage{
assert(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{\dots}{Passthrough to assertthat::assert_that}

\item{env}{environment on which to apply the assertion}

\item{msg}{overrides the error message, if specified.}
}
\description{
Shorter version of assertthat::assert_that
}
\concept{assert}
