% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{consequential}
\alias{consequential}
\title{Checks consequential presence of arguments.}
\usage{
consequential(anchor, dependent)
}
\arguments{
\item{anchor}{The parameter that decides if the other must be
NULL or not.}

\item{dependent}{The parameter that is dependent on anchor being NULL or not}
}
\description{
Check that passes in either of the two cases:
- if the first parameter (anchor) is not NULL, then
  the second parameter (dependent) must also not be NULL, or
- if the first parameter (anchor) is NULL, then
  the second parameter (dependent) must also be NULL

This test is generally useful if one or more parameters in a routine are
required if another parameter is passed, and are therefore consequential
to its presence.
}
\examples{
\dontrun{
# For assertion. This passes if either is satisfied:
# If a is not NULL, then b must also be not NULL.
# If a is NULL, then b must also be NULL.
assertthat::assert_that(
  qscheck::consequential(a, b)
)
# For check
if (
  qscheck::consequential(a, b)
) {}
}

}
\concept{arguments}
