% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{data_frames_same_colnames}
\alias{data_frames_same_colnames}
\title{Checks if two data frames have the exact same column names}
\usage{
data_frames_same_colnames(df1, df2)
}
\arguments{
\item{df1}{the first data frame}

\item{df2}{the first data frame}
}
\description{
Checks if two data frames have the exact same column names
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::data_frames_same_colnames(df1, df2))
# For check
if (qscheck::data_frames_same_colnames(df1, df2)) {}
}

}
