% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj-report.R
\name{failure}
\alias{failure}
\title{Function to create a failure. Returns a qscheck::report object.}
\usage{
failure(reason = NULL, tag = NULL, subreport = NULL)
}
\arguments{
\item{reason}{A string explaining the reason for the failure to the user.}

\item{tag}{A unique tag to identify the failure}

\item{subreport}{if the failure arises from a subcheck, this will contain
the report of that subcheck}
}
\description{
Function to create a failure. Returns a qscheck::report object.
}
