% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_data_frame}
\alias{is_data_frame}
\title{Checks if the passed entity is a data frame.}
\usage{
is_data_frame(
  df,
  exact_rownames = NULL,
  exact_colnames = NULL,
  required_rownames = NULL,
  required_colnames = NULL,
  allow_na = FALSE,
  allow_null = FALSE
)
}
\arguments{
\item{df}{the entity that might be a data frame}

\item{exact_rownames}{vector of strings that, if passed, add the
requirement for the data frame to contain
*exactly* the named rows. Note that if a name
has NULL assigned to it, it counts as not present.}

\item{exact_colnames}{as required_rownames, but for column names.}

\item{required_rownames}{vector of strings that, if passed, add the
requirement for the data frame to contain *at least*
those named rows, respectively. Note that if a name
has NULL assigned to it, it counts as not present.
If exact_rownames is present, this option is ignored.}

\item{required_colnames}{as required_rownames, but for column names}

\item{allow_na}{if TRUE, NA is accepted as a valid value for df.
if FALSE (default) do not accept it.}

\item{allow_null}{if TRUE, NULL is accepted as a valid value.
If FALSE (default) do not accept it.}
}
\description{
Checks if the passed entity is a data frame.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_data_frame(my_parameter))
# For check
if (qscheck::is_data_frame()) {}
}

}
