% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real.R
\name{is_decreasing_vector}
\alias{is_decreasing_vector}
\title{Checks if the passed entity is a vector of decreasing numerical values.}
\usage{
is_decreasing_vector(v, strictly = FALSE, allow_na_values = FALSE)
}
\arguments{
\item{v}{The vector to check}

\item{strictly}{boolean: If TRUE, only strictly decreasing vectors will
satisfy the test.}

\item{allow_na_values}{boolean: If TRUE, NA values are allowed in the vector
and are considered as not existent.
If FALSE, fails for vectors containing NAs.}
}
\description{
Checks if the passed entity is a vector of decreasing numerical values.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_decreasing_vector(my_parameter))
# For check
if (qscheck::is_decreasing_vector(my_parameter)) {}
}

}
\concept{vector}
