% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{is_diagonal_matrix}
\alias{is_diagonal_matrix}
\title{Check if the passed entity is a diagonal matrix. Elements outside of
the diagonal are checked against a given tolerance.}
\usage{
is_diagonal_matrix(
  value,
  exact_dimension = NULL,
  allow_null = FALSE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{value}{The value to check.}

\item{exact_dimension}{If specified, the matrix must have the specified
exact dimension.}

\item{allow_null}{If TRUE, allow NULL as a value.}

\item{tol}{The tolerance to verify if the off-diagonal elements are
zero. Default is sqrt(.Machine$double.eps).}
}
\description{
Check if the passed entity is a diagonal matrix. Elements outside of
the diagonal are checked against a given tolerance.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_diagonal_matrix(value))
# For check
if (qscheck::is_diagonal_matrix(value)) {}
}

}
\concept{matrix}
