% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{is_existing_dir}
\alias{is_existing_dir}
\title{Checks if the path refers to an existing directory.}
\usage{
is_existing_dir(path, allow_null = FALSE)
}
\arguments{
\item{path}{the value to check}

\item{allow_null}{if TRUE, allow the value NULL. If FALSE (default) do not
allow NULL.}
}
\description{
Checks if the path refers to an existing directory.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_existing_dir(f))
# For check
if (qscheck::is_existing_dir(f)) {}
}


}
\concept{filesystem}
