% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{is_existing_file}
\alias{is_existing_file}
\title{Checks if the path refers to an existing file.
Note that this checks for presence of a file. It will fail for directories.}
\usage{
is_existing_file(path, allow_null = FALSE)
}
\arguments{
\item{path}{the path to check}

\item{allow_null}{if TRUE, allow the path to be NULL.
If FALSE (default) do not allow NULL.}
}
\description{
Checks if the path refers to an existing file.
Note that this checks for presence of a file. It will fail for directories.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::file_exists(f))
# For check
if (qscheck::file_exists(f)) {}
}


}
\concept{filesystem}
