% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_factor}
\alias{is_factor}
\title{Check if the passed entity is a factor.}
\usage{
is_factor(value, exact_levels = NULL, exact_length = NULL, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{exact_levels}{vector of strings of the expected levels of the factor.
The factor must contain *exactly* the specified elements.}

\item{exact_length}{integer value. If passed, the factor must have the
*exact* specified length.}

\item{allow_null}{if TRUE, NULL is accepted as a valid value.
If FALSE (default) do not accept it.}
}
\description{
Check if the passed entity is a factor.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_factor(my_parameter))
# For check
if (qscheck::is_factor(my_parameter)) {}
}


}
