% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_function}
\alias{is_function}
\title{Checks if the value is a function}
\usage{
is_function(value, args = NULL, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{args}{a vector of strings. If defined, the function must have
the exact arguments, in the specified order.}

\item{allow_null}{if TRUE, allow the value NULL. If FALSE (default) do not
allow NULL.}
}
\description{
Checks if the value is a function
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_function(f))
# For check
if (qscheck::is_function(f)) {}
}


}
