% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{is_identity_matrix}
\alias{is_identity_matrix}
\title{Check if the passed entity is an identity matrix}
\usage{
is_identity_matrix(
  value,
  exact_dimension = NULL,
  allow_null = FALSE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{value}{the value to check}

\item{exact_dimension}{If specified, the matrix must have the specified
exact dimension}

\item{allow_null}{If TRUE, allow NULL as a value}
}
\description{
Check if the passed entity is an identity matrix
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_identity_matrix(value))
# For check
if (qscheck::is_identity_matrix(value)) {}
}

}
\concept{matrix}
