% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_integer_value}
\alias{is_integer_value}
\title{Checks if the passed value is a single integer value (not type)}
\usage{
is_integer_value(value, min = NULL, max = NULL, allow_null = FALSE)
}
\arguments{
\item{value}{the value to verify}

\item{min}{minimum value to constraint the integer, inclusive}

\item{max}{maximum value to constraint the integer, inclusive}

\item{allow_null}{if TRUE, accepts a null value}
}
\description{
Checks if the passed value is a single integer value (not type)
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_integer_value(my_parameter))
# For check
if (qscheck::is_integer_value(my_parameter)) {}
}

}
