% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{is_list}
\alias{is_list}
\title{Check if the passed entity is a list.}
\usage{
is_list(l, required_names = NULL, exact_length = NULL, allow_null = FALSE)
}
\arguments{
\item{l}{the value to check}

\item{required_names}{vector of strings. If passed, checks if the list
contains *at least* the specified elements.}

\item{exact_length}{integer value. If passed, the list must have the
*exact* specified length}

\item{allow_null}{if TRUE, NULL is accepted as a valid value.
If FALSE (default) do not accept it.}
}
\description{
Check if the passed entity is a list.
}
\details{
Note that a list with names set to NULL are considered not to be present.
This is an R detail, not a qscheck detail.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_list(my_parameter))
# For check
if (qscheck::is_list(my_parameter)) {}
}


}
