% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{is_logical_value}
\alias{is_logical_value}
\title{Check if the passed entity is a single logical value (TRUE or FALSE)}
\usage{
is_logical_value(value, allow_na = FALSE, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{allow_na}{if TRUE, accepts a NA value}

\item{allow_null}{if TRUE, accepts a null value}
}
\description{
Check if the passed entity is a single logical value (TRUE or FALSE)
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_logical_value(value))
# For check
if (qscheck::is_logical_value(value)) {}
}

}
\concept{logical}
