% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{is_lt_value}
\alias{is_lt_value}
\title{Check if the passed entity is a single floating point which is
less than another specified value.}
\usage{
is_lt_value(value, comparator, allow_na = FALSE, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{comparator}{the maximum allowed value, exclusive.}

\item{allow_na}{if true, accept a value that is NA.}

\item{allow_null}{if true, accept a value that is NULL.}
}
\description{
Check if the passed entity is a single floating point which is
less than another specified value.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_lt_value(value, comparator))
# For check
if (qscheck::is_lt_value(value, comparator)) {}
}

}
\concept{real}
