% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{is_matrix}
\alias{is_matrix}
\title{Check if the passed entity is a matrix}
\usage{
is_matrix(
  value,
  exact_num_rows = NULL,
  exact_num_cols = NULL,
  allow_null = FALSE
)
}
\arguments{
\item{value}{the value to check}

\item{exact_num_rows}{If specified, the matrix must have the specified
exact number of rows}

\item{exact_num_cols}{If specified, the matrix must have the specified
exact number of cols}

\item{allow_null}{If TRUE, allow NULL as a value}
}
\description{
Check if the passed entity is a matrix
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_matrix(value))
# For check
if (qscheck::is_matrix(value)) {}
}

}
\concept{matrix}
