% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{is_na_real_value}
\alias{is_na_real_value}
\title{Checks if a value is a single NA of real type}
\usage{
is_na_real_value(value)
}
\arguments{
\item{value}{The value to check}
}
\description{
Allows to check if a given value is a single NA value of type real
}
\examples{
\dontrun{
assertthat::assert_that(
  qscheck::is_na_real_value(value)
)
}

}
\concept{na}
