% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_value.R
\name{is_non_negative_integer_value}
\alias{is_non_negative_integer_value}
\title{Checks if the value is a single non negative integer value (not type)}
\usage{
is_non_negative_integer_value(value, allow_na = FALSE, allow_null = FALSE)
}
\arguments{
\item{value}{the value to verify}

\item{allow_na}{accepts a NA value}

\item{allow_null}{accepts a null value}
}
\description{
Checks if the value is a single non negative integer value (not type)
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_non_negative_integer_value(my_parameter))
# For check
if (qscheck::is_non_negative_integer_value(my_parameter)) {}
}

}
\concept{integer}
