% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_positive_integer_value}
\alias{is_positive_integer_value}
\title{Checks if the value is a single positive integer value (not type)}
\usage{
is_positive_integer_value(value, allow_null = FALSE)
}
\arguments{
\item{value}{the value to verify}

\item{allow_null}{accepts a null value}
}
\description{
Checks if the value is a single positive integer value (not type)
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_positive_integer_value(my_parameter))
# For check
if (qscheck::is_positive_integer_value(my_parameter)) {}
}

}
