% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_r6_class}
\alias{is_r6_class}
\title{Checks if the passed entity is a R6 class.}
\usage{
is_r6_class(value, class_name)
}
\arguments{
\item{value}{the value to check}

\item{class_name}{The name of the class}
}
\description{
Checks if the passed entity is a R6 class.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_r6_class(value, "MyClass"))
# For check
if (qscheck::is_r6_class(value, "MyClass")) {}
}

}
