% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_r6_instance}
\alias{is_r6_instance}
\title{Checks if the passed entity is an instance of a given R6 class name.}
\usage{
is_r6_instance(value, class_name, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{class_name}{The name of the class}

\item{allow_null}{If TRUE, allow the value to be NULL. If FALSE (default)
do not accept NULL.}
}
\description{
Checks if the passed entity is an instance of a given R6 class name.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_r6_instance(value, "MyClass"))
# For check
if (qscheck::is_r6_instance(value, "MyClass")) {}
}


}
