% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_value.R
\name{is_real_value}
\alias{is_real_value}
\title{Check if the passed entity is a single floating point value.
Optionally, specify the minimum and maximum values allowed, either
inclusive or exclusive.}
\usage{
is_real_value(
  value,
  min = NULL,
  max = NULL,
  inclusive_min = TRUE,
  inclusive_max = TRUE,
  allow_na = FALSE,
  allow_null = FALSE
)
}
\arguments{
\item{value}{the value to check}

\item{min}{the minimum allowed value, inclusive or exclusive.}

\item{max}{the maximum allowed value, inclusive or exclusive.}

\item{inclusive_min}{if TRUE (default) the min value is checked inclusive.
If FALSE, the min value will be checked exclusive.}

\item{inclusive_max}{if TRUE (default) the max value is checked inclusive.
If FALSE, the max value will be checked exclusive}

\item{allow_na}{if true, accept a value that is NA.}

\item{allow_null}{if true, accept a value that is NULL.}
}
\description{
Check if the passed entity is a single floating point value.
Optionally, specify the minimum and maximum values allowed, either
inclusive or exclusive.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_real_value(value))
# For check
if (qscheck::is_real_value(value)) {}
}

}
\concept{real}
