% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_vector.R
\name{is_real_vector}
\alias{is_real_vector}
\title{Checks if the passed entity is a vector of reals (numeric).
Note that in R single values are also vectors.}
\usage{
is_real_vector(
  value,
  exact_length = NULL,
  min_length = NULL,
  max_length = NULL,
  allow_na_values = FALSE,
  allow_null = FALSE
)
}
\arguments{
\item{value}{The value to check}

\item{exact_length}{integer. If passed, checks if the vector is of the
*exact* specified length.}

\item{min_length}{integer. If passed, checks if the vector length
is at least the specified length, inclusive.
Note: if exact_length is specified, this parameter is
ignored}

\item{max_length}{integer. If passed, checks if the vector length
is at most the specified length, inclusive
Note: if exact_length is specified, this parameter is
ignored}

\item{allow_na_values}{boolean. If passed allows vectors containing
NAs. The length check is performed including
the NA values. Default FALSE.}

\item{allow_null}{boolean. If TRUE, also accepts a value of NULL.
Default FALSE.}
}
\description{
Checks if the passed entity is a vector of reals (numeric).
Note that in R single values are also vectors.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_real_vector(my_parameter))
# For check
if (qscheck::is_real_vector(my_parameter)) {}
}

}
\concept{real}
