% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_s3_instance}
\alias{is_s3_instance}
\title{Checks if the passed entity is an S3 instance.}
\usage{
is_s3_instance(value, class_name, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{class_name}{the name of the class, as a string.}

\item{allow_null}{if true, accept a value that is NULL.}
}
\description{
Checks if the passed entity is an S3 instance.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(
  qscheck::is_s3_instance(value, "MyClass")
  )
# For check
if (qscheck::is_s3_instance(value, "MyClass")) {}
}

}
