% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{is_s4_instance}
\alias{is_s4_instance}
\title{Checks if the passed entity is an S4 instance of a specified class.}
\usage{
is_s4_instance(value, class_name, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{class_name}{the name of the class, as a string.}

\item{allow_null}{if true, accept a value that is NULL.}
}
\description{
Checks if the passed entity is an S4 instance of a specified class.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(
  qscheck::is_s4_instance(value, "MyClass")
  )
# For check
if (qscheck::is_s4_instance(value, "MyClass")) {}
}

}
\concept{oop}
