% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_satisfying_funcs.R
\name{is_satisfying_funcs}
\alias{is_satisfying_funcs}
\title{Checks if a passed value satisfies one or more tests provided as functions.}
\usage{
is_satisfying_funcs(value, func_or_list)
}
\arguments{
\item{value}{the value to verify}

\item{func_or_list}{a function or (optionally named) list of functions.
The function must accept a single argument (which will receive the
value) and return either a logical (TRUE or FALSE), or an instance
of qscheck::report.}
}
\description{
Checks if a passed value satisfies one or more tests provided as functions.
}
\examples{
\dontrun{
qscheck::assert(
   qscheck::is_satisfying_funcs(
     my_parameter,
     list(
        more_than_five = function(x) {
           # Short form. Returns a logical
           x > 5
        },
        less_than_two = function(x) {
           # expanded form. use success() and failure(reason)
           if (x < 2) {
             return(success())
           } else {
             return(failure("Value must be less than two"))
           }
        }
     )
   )
))
}

}
\concept{general}
