% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{is_square_matrix}
\alias{is_square_matrix}
\title{Check if the passed entity is a square matrix}
\usage{
is_square_matrix(value, exact_dimension = NULL, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{exact_dimension}{If specified, the matrix must have the specified
exact dimension}

\item{allow_null}{If TRUE, allow NULL as a value}
}
\description{
Check if the passed entity is a square matrix
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_square_matrix(value))
# For check
if (qscheck::is_square_matrix(value)) {}
}

}
\concept{matrix}
