% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_value.R
\name{is_string_value}
\alias{is_string_value}
\title{Checks if the passed value is a single string.}
\usage{
is_string_value(
  value,
  allow_empty = TRUE,
  allow_na = FALSE,
  allow_null = FALSE
)
}
\arguments{
\item{value}{the value to verify}

\item{allow_empty}{If TRUE (default) accept an empty string. If FALSE, don't}

\item{allow_na}{If TRUE, accepts a single NA value. If FALSE (default) don't}

\item{allow_null}{If TRUE, accepts a single NULL value.
If FALSE (default) don't}
}
\description{
Checks if the passed value is a single string.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_string_value(my_parameter))
# For check
if (qscheck::is_string_value(my_parameter)) {}
}

}
\concept{string}
