% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{is_value}
\alias{is_value}
\title{Check if the passed entity is a single value, regardless of the type.}
\usage{
is_value(value, allow_na = FALSE, allow_null = FALSE)
}
\arguments{
\item{value}{the value to check}

\item{allow_na}{if true, accept a value that is NA of any type.}

\item{allow_null}{if true, accept a value that is NULL.}
}
\description{
Check if the passed entity is a single value, regardless of the type.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_value(value))
# For check
if (qscheck::is_value(value)) {}
}

}
\concept{value}
