% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_vector}
\alias{is_vector}
\title{Checks if the passed entity is a vector. Note that in R single values
are also vectors.}
\usage{
is_vector(value, exact_length = NULL, min_length = NULL, max_length = NULL)
}
\arguments{
\item{value}{The value to check}

\item{exact_length}{integer. If passed, checks if the vector is of the
*exact* specified length.}

\item{min_length}{integer. If passed, checks if the vector length
is at least the specified length, inclusive.
Note: if exact_length is specified, this parameter is
ignored}

\item{max_length}{integer. If passed, checks if the vector length
is at most the specified length, inclusive
Note: if exact_length is specified, this parameter is
ignored}
}
\description{
Checks if the passed entity is a vector. Note that in R single values
are also vectors.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_vector(my_parameter))
# For check
if (qscheck::is_vector(my_parameter)) {}
}

}
