% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{is_vector_all_na}
\alias{is_vector_all_na}
\title{Checks if the passed vector contains only NAs.}
\usage{
is_vector_all_na(value)
}
\arguments{
\item{value}{The value to check}
}
\description{
Checks if the passed vector contains only NAs.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(
  qscheck::is_vector_all_na(my_parameter)
)
# For check
if (qscheck::is_vector_all_na(my_parameter)) {}
}

}
