% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrixes_can_multiply}
\alias{matrixes_can_multiply}
\title{Check if the passed entities are matrixes that can multiply together}
\usage{
matrixes_can_multiply(m1, m2, result_num_rows = NULL, result_num_cols = NULL)
}
\arguments{
\item{m1}{The first matrix}

\item{m2}{The second matrix}

\item{result_num_rows}{The expected number of rows of the result. Optional.}

\item{result_num_cols}{The expected number of cols of the result. Optional.}
}
\description{
Check if the passed entities are matrixes that can multiply together
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::matrixes_can_multiply(m1, m2))
# For check
if (qscheck::matrixes_can_multiply(m1, m2)) {}
}

}
\concept{matrix}
