% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matrixes_same_rows}
\alias{matrixes_same_rows}
\title{Check if the passed entities are matrixes with the same number of rows}
\usage{
matrixes_same_rows(m1, m2)
}
\arguments{
\item{m1}{The first matrix}

\item{m2}{The second matrix}
}
\description{
Check if the passed entities are matrixes with the same number of rows
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::matrixes_same_rows(m1, m2))
# For check
if (qscheck::matrixes_same_rows(m1, m2)) {}
}

}
\concept{matrix}
