% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{message}
\alias{message}
\title{Returns just the error message or NULL if no error.}
\usage{
message(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{\dots}{Condition to test}

\item{env}{environment on which to apply the assertion}

\item{msg}{overrides the error message, if specified.}
}
\description{
Returns just the error message or NULL if no error.
}
