% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{validate}
\alias{validate}
\title{Returns the error message if fails or TRUE if successful.
Useful in S4 validation.}
\usage{
validate(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{\dots}{Condition to test}

\item{env}{environment on which to apply the assertion}

\item{msg}{overrides the error message, if specified.}
}
\description{
Returns the error message if fails or TRUE if successful.
Useful in S4 validation.
}
\concept{assert}
