% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_allowed_values}
\alias{vector_allowed_values}
\title{Checks if a vector contains values only out of a set of available options.}
\usage{
vector_allowed_values(v, allowed_values)
}
\arguments{
\item{v}{the vector}

\item{allowed_values}{a vector of allowed values}
}
\description{
Checks if a vector contains values only out of a set of available options.
}
\examples{
\dontrun{
# For assertion.
# Will fail if v contains any value that is not either 1, 2 or 3
assertthat::assert_that(qscheck::vector_allowed_values(v, c(1, 2, 3)))
# For check
if (qscheck::vector_allowed_values(v, c(1, 2, 3))) {
}
}

}
