% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_vector.R
\name{vector_values_between}
\alias{vector_values_between}
\title{Checks if the values in the vector are between specified min and max values.}
\usage{
vector_values_between(
  v,
  min,
  max,
  inclusive_min = TRUE,
  inclusive_max = TRUE,
  allow_na = FALSE
)
}
\arguments{
\item{v}{the vector to check}

\item{min}{the minimum allowed value for each vector element, inclusive
or exclusive. -Inf is accepted.}

\item{max}{the maximum allowed value for each vector element, inclusive
or exclusive. Inf is accepted.}

\item{inclusive_min}{if TRUE (default) the min value is checked inclusive.
If FALSE, the min value will be checked exclusive.}

\item{inclusive_max}{if TRUE (default) the max value is checked inclusive.
If FALSE, the max value will be checked exclusive}

\item{allow_na}{If TRUE, accept values that are NA. Default FALSE.}
}
\description{
The interval by default is intended as inclusive [min, max].
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::vector_values_between(vec, 3, 5))
# For check
if (qscheck::vector_values_between(vec, 3, 5)) {}
}

}
\concept{vector}
