% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vectors_disjoint}
\alias{vectors_disjoint}
\title{Checks if two vectors are disjoint.}
\usage{
vectors_disjoint(v1, v2)
}
\arguments{
\item{v1}{the first vector}

\item{v2}{the second vector}
}
\description{
Allows to check if two vectors are disjoint. Two vectors
are disjoint if they are of the same type and none of the elements
present in one can be found in the other.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(
  qscheck::vectors_disjoint(v1, v2)
  )
)
# For check
if (
  qscheck::vectors_disjoint(v1, v2)
  )
) {}
}

}
