% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vectors_same_length}
\alias{vectors_same_length}
\title{Checks if two vectors are of the exact same length.}
\usage{
vectors_same_length(v1, v2)
}
\arguments{
\item{v1}{the first vector}

\item{v2}{the first vector}
}
\description{
Checks if two vectors are of the exact same length.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::vectors_same_length(v1, v2))
# For check
if (qscheck::vectors_same_length(v1, v2)) {}
}

}
\concept{vector}
