% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Shorter version of assertthat::assert_that}
\usage{
assert(...)
}
\arguments{
\item{\dots}{Passthrough to assertthat::assert_that}
}
\description{
Shorter version of assertthat::assert_that
}
