% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_vector.R
\name{is_binary_vector}
\alias{is_binary_vector}
\title{Checks if a vector contains only binary values (0 or 1)}
\usage{
is_binary_vector(
  v,
  allow_na_values = FALSE,
  allow_degenerate = NULL,
  allow_uniform = TRUE
)
}
\arguments{
\item{v}{the vector}

\item{allow_na_values}{if the vector can contain NA values. Default FALSE}

\item{allow_degenerate}{deprecated. Use allow_uniform.}

\item{allow_uniform}{if TRUE (the default), the vector is allowed to contain}
}
\description{
Checks if a vector contains only binary values (0 or 1)
}
\examples{
\dontrun{
# For assertion.
# Will fail if v contains any value that is not either 0 or 1
assertthat::assert_that(qscheck::is_binary_vector(v))
# For check
if (qscheck::is_binary_vector(v)) {}
}

}
\concept{vector}
