% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_logical_value}
\alias{is_logical_value}
\title{Check if the passed entity is a single logical value (TRUE or FALSE)}
\usage{
is_logical_value(value)
}
\arguments{
\item{value}{the value to check}
}
\description{
Check if the passed entity is a single logical value (TRUE or FALSE)
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_logical_value(value))
# For check
if (qscheck::is_logical_value(value)) {}
}

}
