% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_positive_real_value}
\alias{is_positive_real_value}
\title{Check if the passed entity is a single floating point positive value.}
\usage{
is_positive_real_value(value)
}
\arguments{
\item{value}{the value to check}
}
\description{
Check if the passed entity is a single floating point positive value.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(qscheck::is_positive_real_value(value))
# For check
if (qscheck::is_positive_real_value(value)) {}
}

}
