% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mutually_exclusive}
\alias{mutually_exclusive}
\title{Checks if the passed values are mutually exclusive.}
\usage{
mutually_exclusive(...)
}
\arguments{
\item{...}{the various values}
}
\description{
Checks if the passed values are mutually exclusive, that is, if
one of the passed values has a non-NULL value, all the others
_must_ be NULL.
If all options are NULL, this test passes.
Allows to check for multiple cases, and pass the test if any of the cases
is satisfied.
This test is generally useful if two or more parameters in a routine are
technically incompatible with each other, and you want the user to know
the passed parameters in that combination don't make any sense.
}
\examples{
\dontrun{
# For assertion
assertthat::assert_that(
  qscheck::mutually_exclusive(parameter1, parameter2, parameter3)
)
# For check
if (
  qscheck::mutually_exclusive(parameter1, parameter2, parameter3)
) {}
}

}
